event_emitter/requirements.txt
amqp
python-socketio